Rewrite dotfiles management setup
Right now, I'm refactoring this setup to be easier to copy and apply my
configuration across a variety of machines. The bare repository strategy
was, indeed, very interesting, but impractical as it was really hard to
use. Now, I'm going to use GNU stow to manage the symlinks to my files.